+++ src/request/query-user-data.ts
@@ -23,11 +23,3 @@ export const queryUserData = async (cookie?: string): Promise<UserDataType> => {
         },
     }, cookie).then((res) => res.data.data.userStatus);
 };
-
-export const queryFavoriteHash = async (cookie: string): Promise<string | undefined> => {
-    return LcAxios(getUrl("favorites"), { method: "GET" }, cookie).then((res) => {
-        const favorites: any[] = res.data?.favorites?.private_favorites || [];
-        const favorite: any | undefined = favorites.find((item: any) => item.name === "Favorite");
-        return favorite && typeof favorite.id_hash === "string" ? favorite.id_hash : undefined;
-    });
-};